+++ tests/testthat/in/roxygenize-cache-success.R
@@ -1,0 +1,4 @@
+#' Title
+#' 
+#' text
+NULL